phase 2 of issue #330
